+++ source/richedit.c
@@ -69,13 +69,13 @@ HB_FUNC(HWG_RE_SETCHARFORMAT)
   SendMessage(hCtrl, EM_HIDESELECTION, 1, 0);
 
   if (HB_ISARRAY(2)) {
-    ULONG ul, ulLen, ulLen1;
+    uint32_t ul, ulLen, ulLen1;
     PHB_ITEM pArr1;
     pArr = hb_param(2, HB_IT_ARRAY);
-    ulLen = (ULONG)hb_arrayLen(pArr);
+    ulLen = (uint32_t)hb_arrayLen(pArr);
     for (ul = 1; ul <= ulLen; ul++) {
       pArr1 = hb_arrayGetItemPtr(pArr, ul);
-      ulLen1 = (ULONG)hb_arrayLen(pArr1);
+      ulLen1 = (uint32_t)hb_arrayLen(pArr1);
       chrNew.cpMin = hb_arrayGetNL(pArr1, 1) - 1;
       chrNew.cpMax = hb_arrayGetNL(pArr1, 2) - 1;
       SendMessage(hCtrl, EM_EXSETSEL, 0, (LPARAM)&chrNew);
@@ -230,12 +230,12 @@ HB_FUNC(HWG_RE_CHARFROMPOS)
   HWND hCtrl = hwg_par_HWND(1);
   int32_t x = hb_parni(2);
   int32_t y = hb_parni(3);
-  ULONG ul;
+  uint32_t ul;
   POINTL pp;
 
   pp.x = x;
   pp.y = y;
-  ul = (ULONG)SendMessage(hCtrl, EM_CHARFROMPOS, 0, (LPARAM)&pp);
+  ul = (uint32_t)SendMessage(hCtrl, EM_CHARFROMPOS, 0, (LPARAM)&pp);
   hb_retnl(ul);
 }
 
@@ -244,13 +244,13 @@ HB_FUNC(HWG_RE_GETTEXTRANGE)
 {
   HWND hCtrl = hwg_par_HWND(1);
   TEXTRANGE tr;
-  ULONG ul;
+  uint32_t ul;
 
   tr.chrg.cpMin = hb_parnl(2) - 1;
   tr.chrg.cpMax = hb_parnl(3) - 1;
 
   tr.lpstrText = (LPTSTR)hb_xgrab((tr.chrg.cpMax - tr.chrg.cpMin + 2) * sizeof(TCHAR));
-  ul = (ULONG)SendMessage(hCtrl, EM_GETTEXTRANGE, 0, (LPARAM)&tr);
+  ul = (uint32_t)SendMessage(hCtrl, EM_GETTEXTRANGE, 0, (LPARAM)&tr);
   HB_RETSTRLEN(tr.lpstrText, ul);
   hb_xfree(tr.lpstrText);
 }
@@ -260,12 +260,12 @@ HB_FUNC(HWG_RE_GETLINE)
 {
   HWND hCtrl = hwg_par_HWND(1);
   int32_t nLine = hb_parni(2);
-  ULONG uLineIndex = (ULONG)SendMessage(hCtrl, EM_LINEINDEX, (WPARAM)nLine, 0);
-  ULONG ul = (ULONG)SendMessage(hCtrl, EM_LINELENGTH, (WPARAM)uLineIndex, 0);
+  uint32_t uLineIndex = (uint32_t)SendMessage(hCtrl, EM_LINEINDEX, (WPARAM)nLine, 0);
+  uint32_t ul = (uint32_t)SendMessage(hCtrl, EM_LINELENGTH, (WPARAM)uLineIndex, 0);
   LPTSTR lpBuf = (LPTSTR)hb_xgrab((ul + 4) * sizeof(TCHAR));
 
-  *((ULONG *)lpBuf) = ul;
-  ul = (ULONG)SendMessage(hCtrl, EM_GETLINE, nLine, (LPARAM)lpBuf);
+  *((uint32_t *)lpBuf) = ul;
+  ul = (uint32_t)SendMessage(hCtrl, EM_GETLINE, nLine, (LPARAM)lpBuf);
   HB_RETSTRLEN(lpBuf, ul);
   hb_xfree(lpBuf);
 }
@@ -349,7 +349,7 @@ HB_FUNC(HWG_PRINTRTF)
     if (!fSuccess) {
       break;
     }
-    cpMin = (ULONG)SendMessage(hwnd, EM_FORMATRANGE, TRUE, (LPARAM)&fr);
+    cpMin = (uint32_t)SendMessage(hwnd, EM_FORMATRANGE, TRUE, (LPARAM)&fr);
     if (cpMin <= fr.chrg.cpMin) {
       fSuccess = FALSE;
       break;
